fix(sandbox-router): fix lint, add malformed-encoding case, run regression under test-unit
Three more issues from automated review, all fixed here:

1. dev/tools/test-unit runs `go test ./...` without -tags=integration, and
   there is no separate integration presubmit in this repo, so the new
   encoded-slash regression test never actually ran anywhere in CI.
   Dropped the build tag (it only used in-process httptest servers, same
   as the table tests in pathroute_test.go — nothing about it needed a
   real integration environment) and renamed the file to
   pathroute_encoding_test.go to stop implying otherwise. stubLookup
   moved from proxy_integration_test.go to the untagged helpers_test.go
   so it stays shared by both plain and -tags=integration tests.

2. Added a pathroute_test.go case exercising the url.PathUnescape error
   path (malformed "%ZZ" escape) that had no coverage.

3. presubmit-agent-sandbox-lint-go failed on the previous commit:
   golangci-lint's `modernize` flagged three `interface{}` that should be
   `any` in access_log_test.go, and `revive` flagged an unused `r`
   parameter in tracing_test.go. Both fixed.
