+++ sandbox-router/observability/tracing_test.go
@@ -86,7 +86,7 @@ func TestTracingMiddleware_EmptyIdentityWhenUnresolved(t *testing.T) {
 	tp := sdktrace.NewTracerProvider(sdktrace.WithSpanProcessor(recorder))
 	tracer := tp.Tracer("test")
 
-	inner := http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
+	inner := http.HandlerFunc(func(w http.ResponseWriter, _ *http.Request) {
 		w.WriteHeader(http.StatusBadRequest)
 	})
